fix: resolve all clippy warnings, add type aliases
- Add type aliases: FtsResultRow, FtsRawRow, VecResultRow, DrawerMeta (db.rs),
  TripleRow (knowledge_graph.rs)
- Replace all complex inline tuple types with aliases
- #[allow(clippy::too_many_arguments)] on search functions (10-arg signatures)
- #[allow(clippy::needless_lifetimes)] on validate.rs functions (borrow needs explicit 'a)
- Fix manual_strip → strip_prefix in diary topic parser
- Fix manual_contains → .contains() in indexer
- Fix unnecessary_cast in import_palace.rs
- Fix filter_map_identity in import_sessions.rs
- Fix missing_transmute_annotations in main.rs
- clippy -- -D warnings: clean, 144 tests pass

+++ src/import_palace.rs
@@ -59,7 +59,7 @@ pub fn import_palace(db: &Database, source_path: &str) -> Result<(usize, usize)>
                     filed_at,
                 ],
             )?;
-            drawers_imported += changes as usize;
+            drawers_imported += changes;
         }
     }
 
@@ -122,7 +122,7 @@ pub fn import_palace(db: &Database, source_path: &str) -> Result<(usize, usize)>
                     row.source_closet,
                 ],
             )?;
-            triples_imported += changes as usize;
+            triples_imported += changes;
         }
     }
